+++ spec/lib/appsignal/config_spec.rb
@@ -103,7 +103,8 @@ describe Appsignal::Config do
         :enable_host_metrics            => true,
         :enable_minutely_probes         => false,
         :hostname                       => Socket.gethostname,
-        :ca_file_path                   => File.join(resources_dir, "cacert.pem")
+        :ca_file_path                   => File.join(resources_dir, "cacert.pem"),
+        :dns_servers                    => []
       )
     end
 
@@ -292,7 +293,7 @@ describe Appsignal::Config do
       expect(config[:active]).to be_truthy
       expect(config[:name]).to eq "App name"
       expect(config[:debug]).to be_truthy
-      expect(config[:ignore_actions]).to eq ["action1", "action2"]
+      expect(config[:ignore_actions]).to eq %w(action1 action2)
       expect(config[:instrument_net_http]).to be_falsey
       expect(config[:instrument_redis]).to be_falsey
       expect(config[:instrument_sequel]).to be_falsey
@@ -363,13 +364,14 @@ describe Appsignal::Config do
     let(:config) { project_fixture_config(:production) }
     before do
       config[:http_proxy] = "http://localhost"
-      config[:ignore_actions] = ["action1", "action2"]
-      config[:ignore_errors] = ["VerySpecificError", "AnotherError"]
+      config[:ignore_actions] = %w(action1 action2)
+      config[:ignore_errors] = %w(VerySpecificError AnotherError)
       config[:log] = "stdout"
       config[:log_path] = "/tmp"
       config[:hostname] = "app1.local"
       config[:filter_parameters] = %w(password confirm_password)
       config[:running_in_container] = false
+      config[:dns_servers] = ["8.8.8.8", "8.8.4.4"]
       config.write_to_environment
     end
 
@@ -397,7 +399,8 @@ describe Appsignal::Config do
       expect(ENV["_APPSIGNAL_HOSTNAME"]).to                     eq "app1.local"
       expect(ENV["_APPSIGNAL_PROCESS_NAME"]).to                 include "rspec"
       expect(ENV["_APPSIGNAL_CA_FILE_PATH"]).to                 eq File.join(resources_dir, "cacert.pem")
-      expect(ENV).to_not                                        have_key("APPSIGNAL_WORKING_DIR_PATH")
+      expect(ENV["_APPSIGNAL_DNS_SERVERS"]).to                  eq "8.8.8.8,8.8.4.4"
+      expect(ENV).to_not                                        have_key("_APPSIGNAL_WORKING_DIR_PATH")
     end
 
     context "with :working_dir_path" do