+++ spec/lib/appsignal/rack/rails_instrumentation_spec.rb
@@ -46,7 +46,7 @@ if DependencyHelper.rails_present?
       after { middleware.call(env) }
     end
 
-    describe "#call_with_appsignal_monitoring" do
+    describe "#call_with_appsignal_monitoring", :error => false do
       it "should create a transaction" do
         expect(Appsignal::Transaction).to receive(:create).with(
           "1",
@@ -68,8 +68,8 @@ if DependencyHelper.rails_present?
         expect(app).to receive(:call).with(env)
       end
 
-      context "with an error" do
-        let(:error) { VerySpecificError.new }
+      context "with an error", :error => true do
+        let(:error) { VerySpecificError }
         let(:app) do
           double.tap do |d|
             allow(d).to receive(:call).and_raise(error)
@@ -90,7 +90,8 @@ if DependencyHelper.rails_present?
         expect_any_instance_of(Appsignal::Transaction).to receive(:set_http_or_background_queue_start)
       end
 
-      after { middleware.call(env) rescue VerySpecificError }
+      after(:error => false) { middleware.call(env) }
+      after(:error => true) { expect { middleware.call(env) }.to raise_error(VerySpecificError) }
     end
 
     describe "#request_id" do